+++ components/pages/home/expertise/SkillsForCompany.tsx
